+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# For next release
+  * **Markus Freitag**
+    * cmd: add get-totp command
+    * go: update dependencies
+  * **secDre4mer**
+    * pkg/keepassxc/client: implement get-totp
+
+*Not released yet*
+
 # Patch Release v1.5.1 (2023-08-11)
   * **Markus Freitag**
     * cmd: dont initialize the client within a persistentprerun